convert to functional component. Tweak colors for search results. Fix search result header incorrect color. Addressing PR comments from previous search branch. Experimenting trimming approaches.
