+++ _locales/en/messages.json
@@ -464,5 +464,6 @@
   "startedACall": "You called $name$",
   "answeredACall": "Call with $name$",
   "trimDatabase": "Trim Database",
-  "trimDatabaseDescription": "Reduces your message database size to your last 10,000 messages."
+  "trimDatabaseDescription": "Reduces your message database size to your last 10,000 messages.",
+  "trimDatabaseConfirmationBody": "Are you sure you want to delete your $deleteAmount$ oldest received messages?"
 }
